Crate zcash_keys

A crate for Zcash key and address management.

zcash_keys contains Rust structs, traits and functions for creating Zcash spending and viewing keys and addresses.

Feature flags

transparent-inputs — Enables use of transparent key parts and addresses

orchard — Enables use of Orchard key parts and addresses

sapling — Enables use of Sapling key parts and addresses

test-dependencies — Exposes APIs that are useful for testing, such as proptest strategies.

Experimental features

unstable-frost — Exposes unstable APIs that are compatible with FROST key management

unstable — Exposes unstable APIs. Their behaviour may change at any time.

Modules

address

Structs for handling supported address types.

encoding

Encoding and decoding functions for Zcash key and address structs.

keysorchard or sapling or transparent-inputs

Helper functions for managing light client key material.