zcash_keys
Module encoding

Encoding and decoding functions for Zcash key and address structs.

Human-Readable Prefixes (HRPs) for Bech32 encodings are located in the [zcash_primitives::constants] module.

Enums

Bech32DecodeErrorsapling

TransparentCodecError

Traits

AddressCodec

A trait for encoding and decoding Zcash addresses.

Functions

decode_extended_full_viewing_keysapling

Decodes an ExtendedFullViewingKey from a Bech32-encoded string, verifying that it matches the provided human-readable prefix.

decode_extended_spending_keysapling

Decodes an ExtendedSpendingKey from a Bech32-encoded string.

decode_extfvk_with_networksapling

Decodes an ExtendedFullViewingKey and the [NetworkType] that it is intended for use with from a Bech32-encoded string.

decode_payment_addresssapling

Decodes a PaymentAddress from a Bech32-encoded string.

decode_transparent_address

Decodes a TransparentAddress from a Base58Check-encoded string.

encode_extended_full_viewing_keysapling

Writes an ExtendedFullViewingKey as a Bech32-encoded string.

encode_extended_spending_keysapling

Writes an ExtendedSpendingKey as a Bech32-encoded string.

encode_payment_addresssapling

Writes a PaymentAddress as a Bech32-encoded string.

encode_payment_address_psapling

Writes a PaymentAddress as a Bech32-encoded string using the human-readable prefix values defined in the specified network parameters.

encode_transparent_address

Writes a TransparentAddress as a Base58Check-encoded string.

encode_transparent_address_p

Writes a [TransparentAddress] as a Base58Check-encoded string. using the human-readable prefix values defined in the specified network parameters.