zcash_keys::encoding

Trait AddressCodec

pub trait AddressCodec<P>
where
    Self: Sized,
{
    type Error;

    // Required methods
    fn encode(&self, params: &P) -> String;
    fn decode(params: &P, address: &str) -> Result<Self, Self::Error>;
}

A trait for encoding and decoding Zcash addresses.

Required Associated Types

type Error

Required Methods

fn encode(&self, params: &P) -> String

Encode a Zcash address.

Arguments

params - The network the address is to be used on.

fn decode(params: &P, address: &str) -> Result<Self, Self::Error>

Decodes a Zcash address from its string representation.

Arguments

params - The network the address is to be used on.

address - The string representation of the address.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.