zcash_keys

Module keys

Available on crate features orchard or sapling or transparent-inputs only.

Helper functions for managing light client key material.

Modules

saplingsapling

testingtest-dependencies

transparenttransparent-key-encoding

Types and encodings for standalone transparent spending keys.

Structs

ReceiverRequirements

Specification for how a unified address should be generated from a unified viewing key.

UnifiedFullViewingKey

A ZIP 316 unified full viewing key.

UnifiedIncomingViewingKey

A ZIP 316 unified incoming viewing key.

UnifiedSpendingKey

A set of spending keys that are all associated with a single ZIP-0032 account identifier.

Enums

AddressGenerationError

Errors that can occur in the generation of unified addresses.

DecodingError

A type for errors that can occur when decoding keys from their serialized representations.

DerivationError

Eraunstable

A version identifier for the encoding of unified spending keys.

ReceiverRequirement

An enumeration of the ways in which a receiver may be requested to be present in a generated UnifiedAddress.

UnifiedAddressRequest

Specification for how a unified address should be generated from a unified viewing key.